requirements.txt
flask
pytelegrambotapi
python-dotenv
requests
gunicorn
psycopg2-binary    